Lock IWellKnown instance when building initial index
Use double-checked locking to avoid an exception when attempting to add
an item but an item with the same key already exists.

Per the documentation, `SortedDictionary[x] = ...` should never throw if
`x` already exists in the collection, but it's not a thread-safe
collection and we weren't initializing it in a critical section, so all
bets were off.