Fix intermittent exception building property index
In production, I've observed several times exceptions similar to the stack trace
reproduced below where there is an attempt to insert an already extant value
into the property indexer's backing `SortedDictionary` instance.

We added a lock around all `PropertyIndexer` initializations in a previous
release because the easiest explanation was multiple threads attempting to
access the indexer at the same time† (leading to two or more threads trying to
initialize the indexer if this happened to be the first time it was accessed),
but the exception continued to occur.

I'm still not sure how it can happen except if `indexer(entity)` yielded a
duplicate value (which we absolutely should have been handling in the first
place), but in reviewing the code it's quickly apparent that the State
abbreviation "AL" is never repeated.

† This is actually a rather unlikely explanation since each thread would create
a new `_indexer` instance and initialize *that*, so it would only occur if timed
in such a way where one thread T1 assigned `_indexer` then thread T2 replaced
`_indexer` with a new instance before T1 made the to `_indexer.AddToIndex(..)`,
otherwise each would be working on a different instance.

Relevant portion of stack trace from Seq:

System.ArgumentException: An item with the same key has already been added. Key: [AL, NeoSmart.Geographical.UnitedStates+State]
   at System.Collections.Generic.TreeSet`1.AddIfNotPresent(T item)
   at System.Collections.Generic.SortedSet`1.Add(T item)
   at System.Collections.Generic.SortedDictionary`2.set_Item(TKey key, TValue value)
   at NeoSmart.Geographical.PropertyIndexer`2.AddToIndex(IEnumerable`1 entities, Func`2 indexer, Func`2 precondition)
   at NeoSmart.Geographical.WellKnown`1.BuildIndex()
   at NeoSmart.Geographical.WellKnown`1.TryGetValue(String search, T& result)
   at NeoSmart.Geographical.WellKnown`1.Find(String search)